input SDL: Split in top and bottom
Split the SDL input driver in a top and bottom
to enable using it with embedded libCs.
