+++ drivers/input/CMakeLists.txt
@@ -7,5 +7,13 @@ zephyr_library_sources_ifdef(CONFIG_INPUT_FT5336 input_ft5336.c)
 zephyr_library_sources_ifdef(CONFIG_INPUT_GPIO_KEYS input_gpio_keys.c)
 zephyr_library_sources_ifdef(CONFIG_INPUT_GPIO_QDEC input_gpio_qdec.c)
 zephyr_library_sources_ifdef(CONFIG_INPUT_NPCX_KBD input_npcx_kbd.c)
-zephyr_library_sources_ifdef(CONFIG_INPUT_SDL_TOUCH input_sdl_touch.c)
 zephyr_library_sources_ifdef(CONFIG_INPUT_XPT2046 input_xpt2046.c)
+
+if (CONFIG_INPUT_SDL_TOUCH)
+  zephyr_library_sources(input_sdl_touch.c)
+  if (CONFIG_NATIVE_APPLICATION)
+    zephyr_library_sources(input_sdl_touch_bottom.c)
+  else()
+    target_sources(native_simulator INTERFACE input_sdl_touch_bottom.c)
+  endif()
+endif()
